backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.7.4
langchain==0.3.0
langchain-community==0.3.0
langchain-ollama==0.2.0
pgvector==0.3.2
pydantic==2.9.2
python-multipart==0.0.12
Pillow==10.4.0
